feat: auditor flow
