+++ components/layouts/compliance_layout.tsx
@@ -1,28 +1,6 @@
 import React from "react";
 import DashBoardPageLayout from "@components/layouts/dashboard_page_layout";
-
-const SortbarData = [
-  {
-    id: 1,
-    title: "Company",
-    path: "/applications/compliance/company",
-  },
-  {
-    id: 2,
-    title: "Screening",
-    path: "/applications/compliance/screening",
-  },
-  {
-    id: 3,
-    title: "Persons",
-    path: "/applications/compliance/persons",
-  },
-  {
-    id: 4,
-    title: "Documents",
-    path: "/applications/compliance/documents",
-  },
-];
+import { useRouter } from "next/router";
 
 interface ComplianceLayoutProps {
   children: React.ReactNode;
@@ -32,6 +10,30 @@ export default function ComplianceLayout({
   children,
   showexport,
 }: ComplianceLayoutProps) {
+  const router = useRouter();
+  const applicationID = router.query.applicationId;
+  const SortbarData = [
+    {
+      id: 1,
+      title: "Company",
+      path: `/applications/${applicationID}/compliance/company`,
+    },
+    {
+      id: 2,
+      title: "Screening",
+      path: `/applications/${applicationID}/compliance/screening`,
+    },
+    {
+      id: 3,
+      title: "Persons",
+      path: `/applications/${applicationID}/compliance/persons`,
+    },
+    {
+      id: 4,
+      title: "Documents",
+      path: `/applications/${applicationID}/compliance/documents`,
+    },
+  ];
   return (
     <DashBoardPageLayout
       showexport={showexport}
