+++ components/layouts/rs_layout.tsx
@@ -1,18 +1,6 @@
 import React from "react";
 import DashBoardPageLayout from "@components/layouts/dashboard_page_layout";
-
-const SortbarData = [
-  {
-    id: 1,
-    title: "Research",
-    path: "/applications/research/research",
-  },
-  {
-    id: 2,
-    title: "Sentiment",
-    path: "/applications/research/sentiments",
-  },
-];
+import { useRouter } from "next/router";
 
 interface researchLayoutProps {
   children: React.ReactNode;
@@ -22,6 +10,20 @@ export default function ResearchLayout({
   children,
   showexport,
 }: researchLayoutProps) {
+  const router = useRouter();
+  const applicationID = router.query.applicationId;
+  const SortbarData = [
+    {
+      id: 1,
+      title: "Research",
+      path: `/applications/${applicationID}/research/research`,
+    },
+    {
+      id: 2,
+      title: "Sentiment",
+      path: `/applications/${applicationID}/research/sentiments`,
+    },
+  ];
   return (
     <DashBoardPageLayout
       showexport={showexport}
