+++ components/layouts/status_layout.tsx
@@ -1,28 +1,6 @@
 import React from "react";
 import DashBoardPageLayout from "@components/layouts/dashboard_page_layout";
-
-const SortbarData = [
-  {
-    id: 1,
-    title: "Overview",
-    path: "/applications/status/overview",
-  },
-  {
-    id: 2,
-    title: "Audit Trail",
-    path: "/applications/status/audit",
-  },
-  {
-    id: 3,
-    title: "Tasks & Team",
-    path: "/applications/status/tasks_teams",
-  },
-  {
-    id: 4,
-    title: "Customer Comms.",
-    path: "/applications/status/customer_comms",
-  },
-];
+import { useRouter } from "next/router";
 
 interface StatusLayoutProps {
   children: React.ReactNode;
@@ -32,6 +10,30 @@ export default function StatusLayout({
   children,
   showexport,
 }: StatusLayoutProps) {
+  const router = useRouter();
+  const applicationID = router.query.applicationId;
+  const SortbarData = [
+    {
+      id: 1,
+      title: "Overview",
+      path: `/applications/${applicationID}/status/overview`,
+    },
+    {
+      id: 2,
+      title: "Audit Trail",
+      path: `/applications/${applicationID}/status/audit`,
+    },
+    {
+      id: 3,
+      title: "Tasks & Team",
+      path: `/applications/${applicationID}/status/tasks_teams`,
+    },
+    {
+      id: 4,
+      title: "Customer Comms.",
+      path: `/applications/${applicationID}/status/customer_comms`,
+    },
+  ];
   return (
     <DashBoardPageLayout
       showexport={showexport}
