+++ package.json
@@ -30,7 +30,7 @@
     "@react-pdf-viewer/default-layout": "^3.6.0",
     "@react-pdf-viewer/thumbnail": "^3.6.0",
     "@react-pdf-viewer/zoom": "^3.6.0",
-    "@taraldefi/tariala-component-library": "^1.16.5",
+    "@taraldefi/tariala-component-library": "^1.18.0",
     "axios": "^0.27.2",
     "cors": "^2.8.5",
     "express": "^4.18.1",
