+++ pages/applications/[applicationId]/financial/balance/index.tsx
