+++ pages/applications/[applicationId]/financial/cash/index.tsx
