+++ pages/applications/[applicationId]/financial/profit/index.tsx
