+++ pages/applications/[applicationId]/research/research/index.tsx
