+++ pages/applications/[applicationId]/transaction/buyer/index.tsx
