+++ pages/applications/index.tsx
@@ -1,0 +1,134 @@
+import Layout from "@components/layouts/layout";
+import { ApplicationTable } from "@taraldefi/tariala-component-library";
+import { useRouter } from "next/router";
+
+const TableData = [
+  {
+    applicationId: 262116,
+    product: "Product Title",
+    dateFrom: "12.01.2022",
+    dateTo: "12.02.2022",
+    importerId: "SX-936",
+    importerName: "Ullrich Weigel OH GmbH",
+    status: "Pending",
+  },
+  {
+    applicationId: 262117,
+    product: "Product Title",
+    dateFrom: "12.01.2022",
+    dateTo: "12.02.2022",
+    importerId: "FH-509",
+    importerName: "Lohmann Kuhn AG",
+    status: "Reviewing",
+  },
+  {
+    applicationId: 262118,
+    product: "Product Title",
+    dateFrom: "12.01.2022",
+    dateTo: "12.02.2022",
+    importerId: "FH-509",
+    importerName: "Lohmann Kuhn AG",
+    status: "Reviewing",
+  },
+  {
+    applicationId: 262119,
+    product: "Product Title",
+    dateFrom: "12.01.2022",
+    dateTo: "12.02.2022",
+    importerId: "SX-936",
+    importerName: "Ullrich Weigel OH GmbH",
+    status: "Pending",
+  },
+  {
+    applicationId: 262120,
+    product: "Product Title",
+    dateFrom: "12.01.2022",
+    dateTo: "12.02.2022",
+    importerId: "FH-509",
+    importerName: "Lohmann Kuhn AG",
+    status: "Reviewing",
+  },
+  {
+    applicationId: 262121,
+    product: "Product Title",
+    dateFrom: "12.01.2022",
+    dateTo: "12.02.2022",
+    importerId: "FH-509",
+    importerName: "Lohmann Kuhn AG",
+    status: "Reviewing",
+  },
+  {
+    applicationId: 262122,
+    product: "Product Title",
+    dateFrom: "12.01.2022",
+    dateTo: "12.02.2022",
+    importerId: "SX-936",
+    importerName: "Ullrich Weigel OH GmbH",
+    status: "Pending",
+  },
+  {
+    applicationId: 262123,
+    product: "Product Title",
+    dateFrom: "12.01.2022",
+    dateTo: "12.02.2022",
+    importerId: "FH-509",
+    importerName: "Lohmann Kuhn AG",
+    status: "Reviewing",
+  },
+  {
+    applicationId: 262124,
+    product: "Product Title",
+    dateFrom: "12.01.2022",
+    dateTo: "12.02.2022",
+    importerId: "FH-509",
+    importerName: "Lohmann Kuhn AG",
+    status: "Reviewing",
+  },
+  {
+    applicationId: 262125,
+    product: "Product Title",
+    dateFrom: "12.01.2022",
+    dateTo: "12.02.2022",
+    importerId: "SX-936",
+    importerName: "Ullrich Weigel OH GmbH",
+    status: "Pending",
+  },
+  {
+    applicationId: 262126,
+    product: "Product Title",
+    dateFrom: "12.01.2022",
+    dateTo: "12.02.2022",
+    importerId: "FH-509",
+    importerName: "Lohmann Kuhn AG",
+    status: "Reviewing",
+  },
+  {
+    applicationId: 2621127,
+    product: "Product Title",
+    dateFrom: "12.01.2022",
+    dateTo: "12.02.2022",
+    importerId: "FH-509",
+    importerName: "Lohmann Kuhn AG",
+    status: "Reviewing",
+  },
+];
+function Index() {
+  const router = useRouter();
+  const handleRowItemClick = (id: number) => {
+    router.push(`/applications/${id}/summary`);
+  };
+  return (
+    <Layout>
+      <div className="viewbody">
+        <div style={{ padding: "10%", width: "100%" }}>
+          <ApplicationTable
+            onClick={handleRowItemClick}
+            applicationTableData={TableData}
+          ></ApplicationTable>
+        </div>
+      </div>
+    </Layout>
+  );
+}
+
+export default Index;
