+++ styles/financials.scss
@@ -65,6 +65,7 @@
     background-color: #fff;
 
     .contentBox {
+      width: 25%;
       display: flex;
       flex-direction: column;
 
@@ -296,6 +297,7 @@
     background-color: #fff;
 
     .contentBox {
+      width: 25%;
       display: flex;
       flex-direction: column;
 
@@ -547,6 +549,7 @@
     background-color: #fff;
 
     .contentBox {
+      width: 25%;
       display: flex;
       flex-direction: column;
