+++ styles/leftMenu.scss
@@ -5,7 +5,7 @@
   top: 126px;
   border-right: 1px solid #e2e8f0;
   padding-right: 0px;
-  width: 20vw;
+  width: 19vw;
   overflow: scroll;
   .selected {
     font-weight: 600;
