+++ styles/main.scss
@@ -287,7 +287,7 @@
     }
 
     .selected {
-      background-color: #0d8489;
+      background-color: #1ab98b;
       color: white;
       // padding: 7px;
     }
