+++ styles/transactions.scss
@@ -454,7 +454,7 @@
   display: flex;
   gap: 15px;
   flex-direction: row;
-  justify-content: space-between;
+  justify-content: start;
 
   .companyDetails {
     color: #475569;
